Apply the transform matrix of points to packed geo
The Packed Transform Points OTL applies the transformation from the points of the second input by modifying the packed primitives intrinsic attributes. It's very useful for transforming alembic geometry without the need of having to unpack them and store the geometry in the IFD for every frame. This way it only references the static geometry and applies the transformation of the points without writing the geometry every frame.