Group SOP written in VEX
The Group VEX OTL replicates the behavior of the Group SOP from Houdini 15.5. The whole node is written in VEX making the performance faster than the original Group SOP. It also has the option to group primitives by volumes or objects. In Houdini 16 a new collection of more optimized Group SOPs were introduced, so this OTL is not as relevant anymore.