Calculate the transform matrix of non-deforming geo
The Extract Transform OTL calculates the transformations of non-deforming geometry and outputs a point representing every piece with it's corresponding transform. It stores the orientation in the N and Up.

It can extract the transform of every piece specified by an attribute, or by connectivity. It can also check to make sure the points used to calculate the transforms aren't collinear.