Generate acropora corals on the incoming terrain
The Procedural Coral OTL generates acropora corals on the terrain specified on its input. It comes with complete control, from the amount of clusters of corals, to the color, density, length and width of every individual coral and it's subcomponents.

It has an option for working with lines and low res geo while modifying the parameters. The whole setup can then be fused into high resolution geometry using VDB. Finally, it has an option to animate the corals as if they were affected by currents.